contest 2 last 12 problem solve

+++ AUTN_Sheet01/Y_Thelast2digits.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main(){
-long long a,b,c,d;
+int a,b,c,d;
 
 cin>>a>>b>>c>>d;
 long long m = (a*b*c*d)%100;
